app.py
import os
from typing import Annotated, TypedDict, Literal
from dotenv import load_dotenv
from langchain_openai import AzureChatOpenAI
from langchain_core.messages import BaseMessage, HumanMessage, AIMessage
from langgraph.graph import StateGraph, START, END
from langgraph.graph.message import add_messages
from langgraph.checkpoint.memory import MemorySaver
from langchain_core.messages import AIMessage

load_dotenv()

class State(TypedDict):
    messages: Annotated[list[BaseMessage], add_messages]

llm = AzureChatOpenAI(
    azure_deployment=os.getenv("AZURE_OPENAI_DEPLOYMENT_NAME"),
    api_version=os.getenv("AZURE_OPENAI_API_VERSION"),
    temperature=0,
)

# 4. Define Nodes (Requirement 2 & 5)
def chatbot_node(state: State):
    """Standard LLM processing node."""
    response = llm.invoke(state["messages"])
    return {"messages": [response]}

def calculator_node(state: State):
    """Custom node for math tasks with mocked usage_metadata."""
    # ... your math logic ...
    result = "Calculator Node: I detected a math query. Result: 100"
    
    # Return message with usage_metadata to satisfy the monitoring script
    return {
        "messages": [
            AIMessage(
                content=result,
                usage_metadata={
                    "input_tokens": 0, 
                    "output_tokens": 0, 
                    "total_tokens": 0
                }
            )
        ]
    }

# 5. Define Routing Logic (Requirement 5: DAG)
def route_question(state: State) -> Literal["calculator", "chatbot"]:
    """Routes to calculator if keywords are found, else to LLM."""
    user_input = state["messages"][-1].content.lower()
    if any(op in user_input for op in ["+", "-", "*", "/", "calculate"]):
        return "calculator"
    return "chatbot"

# 6. Build the Graph (Requirement 2, 3, 5)
workflow = StateGraph(State)

# Add Nodes
workflow.add_node("chatbot", chatbot_node)
workflow.add_node("calculator", calculator_node)

# Add Edges (The DAG structure)
workflow.add_conditional_edges(START, route_question)
workflow.add_edge("chatbot", END)
workflow.add_edge("calculator", END)

# 7. Add Memory Checkpointer (Requirement 3)
memory = MemorySaver()
app = workflow.compile(checkpointer=memory)

if __name__ == "__main__":
    config = {"configurable": {"thread_id": "1"}}
    user_input = "Hello! Can you remember my name is Aditya?"
    for event in app.stream({"messages": [HumanMessage(content=user_input)]}, config):
        print(event)